Bump up dependencies to fix cves (#11765)

+++ dolphinscheduler-dao/src/test/java/org/apache/dolphinscheduler/dao/mapper/TaskDefinitionLogMapperTest.java
@@ -57,7 +57,7 @@ public class TaskDefinitionLogMapperTest extends BaseDaoTest {
     @Test
     public void testInsert() {
         TaskDefinitionLog taskDefinitionLog = insertOne();
-        Assert.assertNotEquals(taskDefinitionLog.getId(), 0);
+        Assert.assertNotEquals(taskDefinitionLog.getId().intValue(), 0);
     }
 
     @Test
